% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{batch_separation}
\alias{batch_separation}
\title{Compute the batch separation metric}
\usage{
batch_separation(emb, batch)
}
\arguments{
\item{emb}{the embedding of the cells where each row is a cell, and each column is a feature (e.g. principal component)}

\item{batch}{the batch labels of the cells}
}
\value{
Cohen's kappa of the the predicted batch labels with the true batch label on the entire dataset (i.e., training accuracy)
}
\description{
Compute the batch separation metric
}
\examples{
# suppose that iris[["Species"]] are the batch labels
batch_separation(emb = iris[,1:2],batch = iris$Species) # lower batch separation present in dim = 1:2
plot(iris[,1],iris[,2],col = iris[["Species"]])

batch_separation(emb = iris[,3:4],batch = iris$Species) # higher batch separation in dim = 3:4
plot(iris[,3],iris[,4],col = iris[["Species"]])
}
