% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine_sces}
\alias{combine_sces}
\title{Combine several SingleCellExperiment objects}
\usage{
combine_sces(sce_list, cell_type_use = "cell_type1", ...)
}
\arguments{
\item{sce_list}{a list of SingleCellExperiment objects}

\item{cell_type_use}{the name of the colData to be used for cell_type, must be available for each member of the list}

\item{...}{addition parameters passed to the factors}
}
\value{
a list with three objects: \code{combined} the combined count data, \code{cell_type} concatenated cell types, \code{batch} the batch labels
}
\description{
Combine several SingleCellExperiment objects
}
