% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_corgi.R
\name{corgi}
\alias{corgi}
\title{The user friendly version of corgi}
\usage{
corgi(X, Y, run_time, must_have_genes = c(), k = 3, fast = T)
}
\arguments{
\item{X}{gene-by-cell expression matrix for batch 1}

\item{Y}{gene-by-cell expression matrix for batch 2}

\item{run_time}{how long to run for}

\item{must_have_genes}{the set of marker genes that is used in every sampled gene set}
}
\value{
a list of scores
}
\description{
The user friendly version of corgi
}
