% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_corgi.R
\name{corgi_select_genes}
\alias{corgi_select_genes}
\title{Select genes}
\usage{
corgi_select_genes(corgi_output, n_genes)
}
\arguments{
\item{corgi_output}{the output of the corgi algorithm}

\item{n_genes}{how many genes to select}
}
\description{
Select genes
}
