% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_corgi.R
\name{dhillon_emb}
\alias{dhillon_emb}
\title{The embedding as defined in Dhillon, Inderjit S. "Co-clustering documents and words using bipartite spectral graph partitioning." Proceedings of the seventh ACM SIGKDD international conference on Knowledge discovery and data mining. ACM, 2001.}
\usage{
dhillon_emb(X, Y, k)
}
\description{
The embedding as defined in Dhillon, Inderjit S. "Co-clustering documents and words using bipartite spectral graph partitioning." Proceedings of the seventh ACM SIGKDD international conference on Knowledge discovery and data mining. ACM, 2001.
}
