% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_axes_legend}
\alias{get_axes_legend}
\title{Returns an empty plot with just the axes and the axes labels}
\usage{
get_axes_legend(emb_name)
}
\description{
Returns an empty plot with just the axes and the axes labels
}
