% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_color_legend}
\alias{get_color_legend}
\title{Legend for colors in a scatter plot}
\usage{
get_color_legend(cell_type, my_color_palette, ncol = NULL,
  legend.position = "right", legend.title = "Cell type", ...)
}
\arguments{
\item{cell_type}{factor or vector of the batch labels}

\item{my_color_palette}{a vector of colors e.g., \code{c("red","blue","green")} or hex color codes}
}
\description{
Useful for creating common color legend for multiple scatter plots
}
