% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_gene_sets.R
\name{get_compared_gene_sets}
\alias{get_compared_gene_sets}
\title{Return gene sets that is compared against CorGI in the manuscript}
\usage{
get_compared_gene_sets(batch1_top_genes, batch1_name = "Batch1",
  batch2_top_genes, batch2_name = "Batch2", desired_size,
  marker_genes = c())
}
\arguments{
\item{batch1_top_genes}{ordered list of genes (e.g., most highly variable to least)}

\item{batch1_name}{name of batch1 (e.g., if \code{batch1_top_genes} was computed using highly variable genes (HVG), then consider \code{batch1_name <- HVG(Batch1)})}

\item{batch2_top_genes}{see `batch1_top_genes`}

\item{batch2_name}{see `batch1_name`}

\item{desired_size}{how big the output gene_sets should be}

\item{marker_genes}{genes that are included in every gene set of the output. Defaults to the empty set, i.e., no markers provided.}
}
\value{
a list of gene sets, where each item is a gene set of size `desired_size`. The list is ordered as follows:
1. `batch1_top_genes[1:desired_size]`
2. `batch2_top_genes[1:desired_size]`
3. `union(batch1_top_genes[1:x],batch2_top_genes[1:y])` so that `x` and `y` are chosen so that the union consists of `desired_size` elements and `|x-y| <= 1` is true
4. `intersect(batch1_top_genes[1:x],batch2_top_genes[1:y])` so that `x` and `y` are chosen so that the intersection consists of `desired_size` elements and `|x-y| <= 1` is true
}
\description{
Return gene sets that is compared against CorGI in the manuscript
}
