% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_gene_sets.R
\name{get_gene_set_of_size_n}
\alias{get_gene_set_of_size_n}
\title{Get gene set of size \code{n} from arbitrary filter \code{gs_func}}
\usage{
get_gene_set_of_size_n(gs_func, n, LB, UB)
}
\arguments{
\item{gs_func}{monotonic nondecreasing function in one (integer-valued) input such that \code{gs_func(x)} is a gene set (represented by a vector)}

\item{n}{number of genes desired}

\item{LB}{lower bound for the input of \code{gs_func}}

\item{UB}{upper bound for the input of \code{gs_func}}
}
\value{
a gene set (a vector of strings)
}
\description{
Binary search algorithm for finding gene sets of a desired size
}
\section{Requirement}{

Monotonic nondecreasing means that whenever \code{x1 >= x2}, \code{length(gs_func(x1)) >= length(gs_func(x2))}.
\code{length(gs_func(LB)) <= n <= length(gs_func(UB))} must hold
}

