% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_membership_labels_from_set_list}
\alias{get_membership_labels_from_set_list}
\title{Convert a list of sets to a membership vector}
\usage{
get_membership_labels_from_set_list(universe, set_list)
}
\arguments{
\item{set_list}{a named list of sets}
}
\description{
Convert a list of sets to a membership vector
}
