% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_scatterplots}
\alias{get_scatterplots}
\title{Return a list of scatter plots}
\usage{
get_scatterplots(embeddings, batch, cell_type)
}
\arguments{
\item{embeddings}{a list of cell embeddings, i.e., \code{embeddings[[i]]} is an n-by-2 matrix of coordinates where n is the number of cells}

\item{batch}{factor or vector of length n}

\item{cell_type}{factor or vector of length n}
}
\value{
a list of scatter plots corresponding to each embedding in \code{embeddings}
}
\description{
Return a list of scatter plots
}
