% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_shape_legend}
\alias{get_shape_legend}
\title{Legend for shapes in a scatter plot}
\usage{
get_shape_legend(batch, my_shape_palette)
}
\arguments{
\item{batch}{factor or vector of the batch labels}

\item{my_shape_palette}{a vector of numbers corresponding to the shape palette http://www.sthda.com/english/wiki/ggplot2-point-shapes}
}
\description{
Useful for creating common shape legend for multiple scatter plots
}
