% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_manipulators.R
\name{HDG_ranking}
\alias{HDG_ranking}
\title{Highly dropped-out gene selection}
\usage{
HDG_ranking(sce, do_normalize = T)
}
\arguments{
\item{sce}{SingleCellExperiment object to be normalized}

\item{do_normalize}{boolean value of whether \code{sce} should be normalized prior to feature selection. Consider setting this to false if you prefer a different normalization technique.}
}
\value{
ranked list such that most informative genes are at the top of the list
}
\description{
Highly dropped-out gene selection
}
