% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dimensionality_reduction}
\alias{plot_dimensionality_reduction}
\title{Wrapper around qplot for make cell scatter plots}
\usage{
plot_dimensionality_reduction(emb, batch, cell_type)
}
\arguments{
\item{emb}{n-by-2 matrix of cell coordinates, where n is the number of cells}

\item{batch}{factor or vector of length n}

\item{cell_type}{factor or vector of length n}
}
\description{
Wrapper around qplot for make cell scatter plots
}
