% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gene_by_cell_heatmap}
\alias{plot_gene_by_cell_heatmap}
\title{Plot gene-by-cell heatmap of a matrix}
\usage{
plot_gene_by_cell_heatmap(mat, annotation_col, order_cells_by,
  annotation_row = NULL, cluster_rows = T, cutree_rows = 7,
  annotate_gene_clusters = T, annotation_colors = list(),
  cluster_cols = F, ...)
}
\arguments{
\item{mat}{gene (row) by cell (column) matrix}

\item{annotation_col}{a data frame of meta data for the cells}

\item{order_cells_by}{which column of annotation_col to use}

\item{annotation_row}{a data frame of meta data for the genes}

\item{cluster_rows}{whether or not to cluster the rows, see pheatmap}

\item{cutree_rows}{number of clusters to be returned by the hiearchical clustering}

\item{annotate_gene_clusters}{boolean value of whether the gene cluster labels should be shown in the row annotation. This is because you might want to cluster the genes but not show the colors.}

\item{annotation_colors}{a list of color palettes (color hex codes), you don't have to supply all or even any colors. By default, the rainbow palette is applied to discrete color scales. For continuous data, black/white gradient is applied.}
}
\description{
Plot gene-by-cell heatmap of a matrix
}
