% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_two_dfs}
\alias{plot_two_dfs}
\title{Plot scatterplots for each pair of columns in df1 and df2}
\usage{
plot_two_dfs(df1, df2)
}
\arguments{
\item{df1}{the first dataframe}

\item{df2}{the second dataframe, rownames(df1) == rownames(df2) must hold}
}
\description{
Code by Ben Bolker, see this \href{https://stackoverflow.com/a/40455040/636276}{stackoverflow answer}
}
\examples{
plot_two_dfs(iris[,1:2],iris[,3:4])
}
\author{
Ben Bolker
}
