% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_venn_diagram}
\alias{plot_venn_diagram}
\title{Plot a venn diagram}
\usage{
plot_venn_diagram(my_sets)
}
\arguments{
\item{my_sets}{a named list of three sets}
}
\description{
Code obtained from
https://scriptsandstatistics.wordpress.com/2018/04/26/how-to-plot-venn-diagrams-using-r-ggplot2-and-ggforce/
}
\examples{
plot_venn_diagram(list(A = 1:5, B = 3:7, C = 5:9))
}
