% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_corgi.R
\name{run_corgi}
\alias{run_corgi}
\title{Corgi feature selection}
\usage{
run_corgi(X, Y, n_phases = 6, time_per_phase = 10 * 60,
  must_have_genes = c(), n_singular_gaps = 3, n_genes_sample = NULL,
  n_cells_X_sample = NULL, n_cells_Y_sample = NULL)
}
\arguments{
\item{X}{gene-by-cell expression matrix for batch 1}

\item{Y}{gene-by-cell expression matrix for batch 2}

\item{n_phases}{how many phases of successive halving to do}

\item{time_per_phase}{how long each phase is (in seconds)}

\item{must_have_genes}{the set of marker genes that is used in every sampled gene set}

\item{n_singular_gaps}{how many singular value gaps to examine}

\item{n_genes_sample}{what size gene set to use}

\item{n_cells_X_sample}{how many cells to sample from X}

\item{n_cells_Y_sample}{how many cells to sample from Y}
}
\value{
a list of list of scores
}
\description{
Corgi feature selection
}
