% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diagnostics.R
\name{run_diagnostics}
\alias{run_diagnostics}
\title{The inner loop of corgi but with more time}
\usage{
run_diagnostics(X, Y, run_time, record_time_interval, genes_use = NULL,
  must_have_genes = c(), n_singular_gaps = 3, n_genes_sample = NULL,
  n_cells_X_sample = NULL, n_cells_Y_sample = NULL)
}
\arguments{
\item{X}{gene-by-cell expression matrix for batch 1}

\item{Y}{gene-by-cell expression matrix for batch 2}

\item{run_time}{how long to run for}

\item{record_time_interval}{how long between each record}

\item{must_have_genes}{the set of marker genes that is used in every sampled gene set}

\item{n_singular_gaps}{how many singular value gaps to examine}

\item{n_genes_sample}{what size gene set to use}

\item{n_cells_X_sample}{how many cells to sample from X}

\item{n_cells_Y_sample}{how many cells to sample from Y}

\item{gene_use}{the set of genes used for sampling}
}
\value{
a list of scores
}
\description{
The inner loop of corgi but with more time
}
