% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{run_mapping_accuracy_comparison}
\alias{run_mapping_accuracy_comparison}
\title{Run scmap over a range of parameters, over multiple gene sets}
\usage{
run_mapping_accuracy_comparison(query, reference, gene_sets,
  method_use = "scmapCluster")
}
\arguments{
\item{query}{a SingleCellExperiment object for the query dataset}

\item{reference}{a SingleCellExperiment object for the reference dataset}

\item{gene_sets}{a named list of gene sets}

\item{method_use}{a string denoting whether to use scmapCluster or scmapCell}
}
\description{
Run scmap over a range of parameters, over multiple gene sets
}
