% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_par_fs.R
\name{run_par_fs}
\alias{run_par_fs}
\title{The inner loop of corgi}
\usage{
run_par_fs(X, Y, run_time, fs, fs_var, genes_use = NULL,
  must_have_genes = c(), n_genes_sample = NULL,
  n_cells_X_sample = NULL, n_cells_Y_sample = NULL,
  n_cores_use = NULL)
}
\arguments{
\item{X}{gene-by-cell expression matrix for batch 1}

\item{Y}{gene-by-cell expression matrix for batch 2}

\item{run_time}{how long to run for}

\item{fs}{a list of scoring functions f(X[F, ], Y[F, ]) -> real number}

\item{fs_var}{global variables, to save time}

\item{must_have_genes}{the set of marker genes that is used in every sampled gene set}

\item{n_genes_sample}{what size gene set to use}

\item{n_cells_X_sample}{how many cells to sample from X}

\item{n_cells_Y_sample}{how many cells to sample from Y}

\item{gene_use}{the set of genes used for sampling}
}
\value{
a list of scores
}
\description{
The inner loop of corgi
}
