% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{run_scmap}
\alias{run_scmap}
\title{Wrapper around the scmapCluster function from scmap}
\usage{
run_scmap(query, ref, gene_set, threshold, ...)
}
\arguments{
\item{query}{the query dataset (SingleCellExperiment object)}

\item{ref}{the reference dataset (SingleCellExperiment object)}

\item{gene_set}{a list of characters representing the genes}

\item{threshold}{the threshold parameter used in scmapCluster, between 0 and 1}
}
\value{
confusion matrix of the true labels in `ref` and the `scmapCluster` predicted labels
}
\description{
Wrapper around the scmapCluster function from scmap
}
