% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{run_scmapCluster}
\alias{run_scmapCluster}
\title{Wrapper for \code{scmap::scmapCluster}}
\usage{
run_scmapCluster(query, ref, gene_set, param, ...)
}
\arguments{
\item{query}{the query dataset (SingleCellExperiment object)}

\item{ref}{the reference dataset (SingleCellExperiment object)}

\item{gene_set}{a list of characters representing the genes}

\item{...}{additional parameters passed on to \code{caret::confusionMatrix}. For example \code{dnn = c("Prediction","Truth")}.}

\item{threshold}{the threshold parameter used in scmapCluster, between 0 and 1}
}
\value{
confusion matrix of the true labels in `ref` and the `scmapCluster` predicted labels
}
\description{
Wrapper for \code{scmap::scmapCluster}
}
