% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sample_cells}
\alias{sample_cells}
\title{Subsample cells random}
\usage{
sample_cells(object, n)
}
\arguments{
\item{object}{an object that is column-subsettable, e.g., a matrix or a SingleCellExperiment object}

\item{n}{how many cells do you want}
}
\description{
Subsample cells random
}
