% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_top_corgi_genes.R
\name{select_top_corgi_genes}
\alias{select_top_corgi_genes}
\title{Get the gene set after running corgi}
\usage{
select_top_corgi_genes(corgi_output, n, phase_use = NULL,
  SVG_use = NULL)
}
\arguments{
\item{corgi_output}{the output of run_corgi}

\item{n}{number of genes in the returned gene set}

\item{phase_use}{which phase of the corgi_output to use, defaults to the last phase}

\item{SVG_use}{which singular value gap (SVG) to use, defaults to the last SVG}
}
\value{
gene set
}
\description{
Get the gene set after running corgi
}
\examples{
data("res_list_liver_bud")
corgi_gene_set <- select_top_corgi_genes(corgi_output_toy_example,10)
corgi_gene_set
}
