% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_CCA_GeneLoading.R
\name{Seurat_CCA_GeneLoading}
\alias{Seurat_CCA_GeneLoading}
\title{Wrapper around Seurat's CCA to extract gene loadings}
\usage{
Seurat_CCA_GeneLoading(mat1, mat2, k = 3)
}
\arguments{
\item{mat1}{batch 1 expression matrix, should be raw}

\item{mat2}{same as above for batch 2}

\item{k}{how many CCA components to return, defaults to k}
}
\value{
gene-by-CCA# matrix where # ranges from 1 to k
}
\description{
Wrapper around Seurat's CCA to extract gene loadings
}
