% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{spearman_rho_mds}
\alias{spearman_rho_mds}
\title{Multidimensional scaling on Spearman rho correlation distance}
\usage{
spearman_rho_mds(X, k = 2)
}
\arguments{
\item{X}{the input matrix, where each column is a sample (cell) and each row is a feature (gene)}

\item{k}{number of latent dimensions, default to 2}
}
\description{
Multidimensional scaling on Spearman rho correlation distance
}
\examples{
mds <- spearman_rho_mds(t(as.matrix(mtcars)))
plot(mds)
text(mds, row.names(mtcars), cex=0.6, pos=4, col="red")
}
