% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_manipulators.R
\name{top_n_genes}
\alias{top_n_genes}
\title{Return the top genes according to a named vector.}
\usage{
top_n_genes(x, n)
}
\arguments{
\item{x}{named numeric vector where the names are the gene names and the numbers are gene scores}

\item{n}{how many genes to return}
}
\description{
Return the top genes according to a named vector.
}
