% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{batch_mixing}
\alias{batch_mixing}
\title{Compute the batch mixing metric}
\usage{
batch_mixing(emb, Batch)
}
\arguments{
\item{emb}{the embedding of the cells where each row is a cell}

\item{Batch}{the batch labels of the cells}
}
\description{
Compute the batch mixing metric
}
