% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{batch_separation}
\alias{batch_separation}
\title{Compute the batch separtion metric}
\usage{
batch_separation(emb, batch)
}
\arguments{
\item{emb}{the embedding of the cells where each row is a cell, and each column is a feature (e.g. principal component)}

\item{batch}{the batch labels of the cells}
}
\value{
Cohen's kappa of the the predicted batch labels with the true batch label on the entire dataset (i.e., training accuracy)
}
\description{
Compute the batch separtion metric
}
