% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_color_legend}
\alias{get_color_legend}
\title{Empty plot with just the axes and labels}
\usage{
get_color_legend(cell_type, my_color_palette, ncol = NULL,
  legend.position = "right", legend.title = "Cell type", ...)
}
\arguments{
\item{emb_name}{name of the embedding, for instance \code{emb_name = "PC"}}
}
\description{
Useful for creating axes legend for multiple scatter plots with the same axes names
}
