% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_gene_sets.R, R/gene_set_manipulators.R
\name{get_compared_gene_sets}
\alias{get_compared_gene_sets}
\title{Gene sets compared against CorGI in the manuscript}
\usage{
get_compared_gene_sets(batch1_top_genes, batch1_name = "Batch1",
  batch2_top_genes, batch2_name = "Batch2", desired_size,
  marker_genes = c())

get_compared_gene_sets(batch1_top_genes, batch1_name = "Batch1",
  batch2_top_genes, batch2_name = "Batch2", desired_size,
  marker_genes = c())
}
\arguments{
\item{batch1_top_genes}{ordered list of genes (e.g., most highly variable to least)}

\item{batch1_name}{name of batch1 (e.g., if \code{batch1_top_genes} was computed using highly variable genes (HVG), then consider \code{batch1_name <- HVG(Batch1)})}

\item{batch2_top_genes}{see \code{batch1_top_genes}}

\item{batch2_name}{see \code{batch1_name}}

\item{desired_size}{how big the output gene_sets should be}

\item{marker_genes}{genes that are included in every gene set of the output. Defaults to the empty set, i.e., no markers provided.}

\item{batch1_top_genes}{ordered list of genes (e.g., most highly variable to least)}

\item{batch1_name}{name of batch1 (e.g., if \code{batch1_top_genes} was computed using highly variable genes (HVG), then consider \code{batch1_name <- HVG(Batch1)})}

\item{batch2_top_genes}{see \code{batch1_top_genes}}

\item{batch2_name}{see \code{batch1_name}}

\item{desired_size}{how big the output gene_sets should be}

\item{marker_genes}{genes that are included in every gene set of the output. Defaults to the empty set, i.e., no markers provided.}
}
\value{
a list of gene sets, where each item is a gene set of size \code{desired_size}. The list is ordered as follows:
\itemize{
  \item \code{batch1_top_genes[1:desired_size]}
  \item \code{batch2_top_genes[1:desired_size]}
  \item \code{union(batch1_top_genes[1:x], batch2_top_genes[1:y])} where \code{x} and \code{y} satisfies \code{|x-y| <= 1}
  \item \code{intersect(batch1_top_genes[1:x],batch2_top_genes[1:y])} where \code{x} and \code{y} satisfies \code{|x-y| <= 1}
}
  If \code{marker_genes} is nonempty, then all four gene sets above include \code{marker_genes}.

a list of gene sets, where each item is a gene set of size \code{desired_size}. The list is ordered as follows:
\itemize{
  \item \code{batch1_top_genes[1:desired_size]}
  \item \code{batch2_top_genes[1:desired_size]}
  \item \code{union(batch1_top_genes[1:x], batch2_top_genes[1:y])} where \code{x} and \code{y} satisfies \code{|x-y| <= 1}
  \item \code{intersect(batch1_top_genes[1:x],batch2_top_genes[1:y])} where \code{x} and \code{y} satisfies \code{|x-y| <= 1}
}
  If \code{marker_genes} is nonempty, then all four gene sets above include \code{marker_genes}.
}
\description{
Gene sets compared against CorGI in the manuscript

Gene sets compared against CorGI in the manuscript
}
\examples{
batch1_top_genes <- c(paste0("g",1:4),paste0("g",7:10))
batch2_top_genes <- paste0("g",3:9)
marker_genes <- "g0"
get_compared_gene_sets(batch1_top_genes,"RankedList1", batch2_top_genes,"RankedList2", 6, marker_genes)
batch1_top_genes <- c(paste0("g",1:4),paste0("g",7:10))
batch2_top_genes <- paste0("g",3:9)
marker_genes <- "g0"
get_compared_gene_sets(batch1_top_genes,"RankedList1", batch2_top_genes,"RankedList2", 6, marker_genes)
}
