% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{run_mapping_accuracy_comparison}
\alias{run_mapping_accuracy_comparison}
\title{Run scmap over a range of parameters}
\usage{
run_mapping_accuracy_comparison(query, reference, gene_sets)
}
\description{
Run scmap over a range of parameters
}
